consolidate some transaction types
