+++ features/borrow/open/pipes/openVaultTransactions.ts
@@ -1,5 +1,4 @@
 import { TxStatus } from '@oasisdex/transactions'
-import { BigNumber } from 'bignumber.js'
 import { open, OpenData } from 'blockchain/calls/proxyActions'
 import { TxMetaKind } from 'blockchain/calls/txMeta'
 import { AddGasEstimationFunction, TxHelpers } from 'components/AppContext'
@@ -10,62 +9,8 @@ import { transactionToX } from 'helpers/form'
 import { zero } from 'helpers/zero'
 import { Observable, of } from 'rxjs'
 
-import { TxError } from '../../../../helpers/types'
 import { parseVaultIdFromReceiptLogs } from '../../../shared/transactions'
 import { OpenVaultChange, OpenVaultState } from './openVault'
-
-type ProxyChange =
-  | {
-      kind: 'proxyWaitingForApproval'
-    }
-  | {
-      kind: 'proxyInProgress'
-      proxyTxHash: string
-    }
-  | {
-      kind: 'proxyFailure'
-      txError?: TxError
-    }
-  | {
-      kind: 'proxyConfirming'
-      proxyConfirmations?: number
-    }
-  | {
-      kind: 'proxySuccess'
-      proxyAddress: string
-    }
-
-type AllowanceChange =
-  | { kind: 'allowanceWaitingForApproval' }
-  | {
-      kind: 'allowanceInProgress'
-      allowanceTxHash: string
-    }
-  | {
-      kind: 'allowanceFailure'
-      txError?: TxError
-    }
-  | {
-      kind: 'allowanceSuccess'
-      allowance: BigNumber
-    }
-
-type OpenChange =
-  | { kind: 'txWaitingForApproval' }
-  | {
-      kind: 'txInProgress'
-      openTxHash: string
-    }
-  | {
-      kind: 'txFailure'
-      txError?: TxError
-    }
-  | {
-      kind: 'txSuccess'
-      id: BigNumber
-    }
-
-export type OpenVaultTransactionChange = ProxyChange | AllowanceChange | OpenChange
 
 export function applyOpenVaultTransaction(
   change: OpenVaultChange,
